refactor: rename custom `random` module
also update all internal calls to module
