+++ neurogym/envs/contextdecisionmaking.py
@@ -52,7 +52,7 @@ class SingleContextDecisionMaking(ngym.TrialEnv):
             "fixation": 300,
             # 'target': 350, # noqa: ERA001
             "stimulus": 750,
-            "delay": ngym.random.TruncExp(600, 300, 3000),
+            "delay": ngym.ngym_random.TruncExp(600, 300, 3000),
             "decision": 100,
         }
         if timing:
@@ -173,7 +173,7 @@ class ContextDecisionMaking(ngym.TrialEnv):
             "fixation": 300,
             # 'target': 350, # noqa: ERA001
             "stimulus": 750,
-            "delay": ngym.random.TruncExp(600, 300, 3000),
+            "delay": ngym.ngym_random.TruncExp(600, 300, 3000),
             "decision": 100,
         }
         if timing:
