+++ neurogym/envs/contrib/changingenvironment.py
@@ -57,7 +57,7 @@ class ChangingEnvironment(ngym.TrialEnv):
 
         self.timing = {
             "fixation": 500,
-            "stimulus": ngym.random.TruncExp(1000, 500, 1500),
+            "stimulus": ngym.ngym_random.TruncExp(1000, 500, 1500),
             "decision": 500,
         }
         if timing:
