+++ neurogym/envs/detection.py
@@ -67,7 +67,7 @@ class Detection(ngym.TrialEnv):
 
         self.timing = {
             "fixation": 500,
-            "stimulus": ngym.random.TruncExp(1000, 500, 1500),
+            "stimulus": ngym.ngym_random.TruncExp(1000, 500, 1500),
         }
         if timing:
             self.timing.update(timing)
