+++ neurogym/envs/readysetgo.py
@@ -242,8 +242,8 @@ class OneTwoThreeGo(ngym.TrialEnv):
             self.rewards.update(rewards)
 
         self.timing = {
-            "fixation": ngym.random.TruncExp(400, 100, 800),
-            "target": ngym.random.TruncExp(1000, 500, 1500),
+            "fixation": ngym.ngym_random.TruncExp(400, 100, 800),
+            "target": ngym.ngym_random.TruncExp(1000, 500, 1500),
             "s1": 100,
             "interval1": (600, 700, 800, 900, 1000),
             "s2": 100,
