+++ neurogym/utils/ngym_random.py
